requirements.txt
dj-database-url==0.5.0
Django==2.2
django-cors-headers==3.2.1
django-heroku==0.3.1
djangorestframework==3.9.2
gunicorn==20.0.4
numpy==1.18.1
opencv-python==4.1.2.30
Pillow==7.0.0
psycopg2==2.8.4
pytz==2019.3
pyzbar==0.1.8
sqlparse==0.3.0
whitenoise==5.0.1
